+++ src/model/common/comtu.h
@@ -7,9 +7,9 @@ class ComTU
 public:
     ComTU();
     ADD_CLASS_FIELD(QVector<ComTU*>, apcTUs, getTUs, setTUs)
-    ADD_CLASS_FIELD(int, iX, getX, setX)                                        ///< X Position in frame
-    ADD_CLASS_FIELD(int, iY, getY, setY)                                        ///< Y Position in frame
-    ADD_CLASS_FIELD(int, iSize, getSize, setSize)                               ///< CU Size
+    ADD_CLASS_FIELD(short, iX, getX, setX)                                        ///< X Position in frame
+    ADD_CLASS_FIELD(short, iY, getY, setY)                                        ///< Y Position in frame
+    ADD_CLASS_FIELD(short, iSize, getSize, setSize)                               ///< CU Size
 
 };
